mental_health_chatbot/requirements.txt
streamlit==1.37.1
pandas==2.2.2
matplotlib==3.9.0
python-dotenv==1.0.1
sqlalchemy==2.0.32
